+++ conf/evolutions/default/76.sql
@@ -18,7 +18,7 @@ CREATE TABLE agentactivity (
   geo_address   TEXT,
   origin_price   DOUBLE,
   price   DOUBLE,
-  priority   INT(4) NOT NULL DEFAULT 0,
+  priority   BIGINT NOT NULL DEFAULT 0,
   publish_status   INT(4) DEFAULT 0,
   published_at BIGINT,
   reject_reason TEXT,
